% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dog_function.R
\name{clean_dog}
\alias{clean_dog}
\title{clean dog}
\usage{
clean_dog(dog_data)
}
\arguments{
\item{dog_data}{the data frame of dog information}
}
\description{
Cleans county dog data
}
\examples{
clean_dog(dog_data)
}
\keyword{clean}
