% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_poodle_frame_function.R
\name{make_poodle_frame}
\alias{make_poodle_frame}
\title{make poodle frame}
\usage{
make_poodle_frame(clean_dog_data)
}
\arguments{
\item{clean_dog_data}{the cleaned data, an output from the clean dog function}
}
\description{
Removes all dogs that aren't poodles from the data frame, renames rows and columns to fit this new data frame, adds a poodle type column in place of breed column
}
\examples{
make_poodle_frame(clean_dog_data)
}
\keyword{poodle,}
\keyword{rename}
