% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_poodle_percent_function.R
\name{make_poodle_percent}
\alias{make_poodle_percent}
\title{make poodle percent}
\usage{
make_poodle_percent(clean_dog_data, poodle_data)
}
\arguments{
\item{clean_dog_data}{the cleaned data frame, an output from the clean dog function}

\item{poodle_data}{the poodle data frame, an output from the make poodle frame function}
}
\description{
makes a data frame of each zip code and poodles as a percent of total dogs
}
\examples{
make_poodle_percent (clean_dog_data, poodle_data)
}
\keyword{percent}
